backend/requirements.txt
fastapi
uvicorn
redis
python-jose[cryptography]
passlib
python-multipart
pydantic-settings
httpx
openai
google-auth-oauthlib
google-api-python-client
aiosmtplib
faiss-cpu
numpy
email-validator
motor
beanie
